S3c: collapse the fire bridge (the CLOSER of S3) + the unit-state determinism digest
The LAST float bridge in the sim FIELD path is gone. After S3c the whole field
path (water + S2 atmosphere/wave/wind/smoke/gas group + fire + temperature) is
integer Q16.16; the only float left in the synced path is the Q2-fenced Python
combat HP math (now digested + watched), the render/cosmetic boundary, and the
documented `dt` boundary cast.

Bridge collapse (plan §S3c):
- temperature_solver.{h,cpp}: the `atmosphere` arg float* -> const int32_t*; the
  vacuum-exposure threshold `atmosphere[n] < o2_vacuum_thresh` is now a Q16.16
  integer compare against `quantize(o2_vacuum_thresh)` (computed once per step,
  the load/boundary cast). The temperature TU is now FULLY INTEGER (its last
  float input gone). Bindings: TemperatureSolver.step atmosphere arg -> int32.
- physics_engine.cpp::step_tail: removed the atm_f_ dequantize loop that fed the
  temperature pass; temperature reads the POST-fire-plume int32 atmosphere
  directly. step_tail has NO float bridge left in the fire/temperature path.
- physics_engine.h: deleted the DEAD scratch wind_x_f_/wind_y_f_/fire_f_ (fire
  went integer in S3b). atm_f_ + wave_p_f_ SURVIVE — they are the SEPARATE,
  documented S2c water head/ripple bridge inside step_water (k_p·(atm+wave_p)
  reads atmosphere as float), retired by a later water/atmosphere unify, NOT S3.

No-float ratchet (plan §S3c CI ratchet): fire_simulation.cpp +
temperature_solver.cpp are now MIGRATED (integer end-to-end). Re-baselined and
added test_migrated_tus_at_documented_floor — a HARD gate pinning these two TUs
at their DOCUMENTED-EXCEPTION floor (the `float dt` boundary cast, FireParams
float members, the load-time quantize/make_recip double precompute, render/glow).
A later patch adding real per-cell float to either TU trips CI.

Unit-state determinism digest (Q5 — HOLD the Python HP fence, ADD the digest):
S3 closes the fire->heat->damage->kill loop, so the float-HP step is now
exercisable. New test_s3c_unit_state_digest.py drives an ignite->fire->
unit-damage->KILL scenario (a held fire deterministically drains a low-HP marine
to death) and asserts the synced unit-state digest (per-unit HP/life/faction/
position/footprint + the hit/kill EVENT stream, defined once in
field_ab_harness.SYNCED_UNIT_FIELDS/_capture_unit_state) is bit-identical
run-to-run, the kill lands on the same tick, and the scenario actually kills
(non-vacuous). The field harness __main__ now also asserts the per-tick unit
digest bit-identity (S3c block).

Gates: P1 self-match tol=0.0 with ZERO float bridges in the sim field path; P2
atmosphere conservation green (the plume deposit is the intended source
exception); the unit-state digest bit-identical across the fire->kill scenario;
full suite green (369); main_smoke --auto exits 0; cross-config self-match
(ticks_per_second + fire params). Temperature tests updated to pass int32 Q16.16
atmosphere. Built cpp/build_vs2022, copied the .pyd into cpp/build/Release.

+++ cpp/src/bindings.cpp
@@ -262,7 +262,7 @@ PYBIND11_MODULE(breach_physics, m) {
                         py::array_t<int32_t> face_shift,
                         py::array_t<bool>    solid,
                         py::array_t<bool>    is_vacuum,
-                        py::array_t<float>   atmosphere) {
+                        py::array_t<int32_t> atmosphere) {   // S3c: Q16.16 int32 (was float)
             auto [temp, h, w]     = get_2d(temperature);
             auto [hp, h2, w2]     = get_2d_const(heat);
             auto [shift, h3, w3]  = get_2d_const(heat_inv_shift);
